+++ demo_menu.gd
@@ -13,12 +13,14 @@ var _current_editor_showed: Control = null
 @onready var _inspector_action_example_button: Button = %InspectorActionButton
 @onready var _viewer_only_example_button: Button = %ViewerOnlyButton
 @onready var _pretty_viewer_example_button: Button = %PrettyViewerButton
+@onready var _interactive_viewer_example_button: Button = %InteractiveViewerButton
 @onready var _demo_mode_panel: Container = %DemoModeESCPanel
 @onready var _minimal_graph_editor_scene: PackedScene = preload("res://addons/custom_graph_editor/examples/minimal/minimal_graph_editor.tscn")
 @onready var _location_map_graph_editor_scene: PackedScene = preload("res://addons/custom_graph_editor/examples/location_map/location_map_editor.tscn")
 @onready var _inspector_action_graph_editor_scene: PackedScene = preload("res://addons/custom_graph_editor/examples/inspector_actions/inspector_actions_editor.tscn")
 @onready var _viewer_only_demo_scene: PackedScene = preload("res://addons/custom_graph_editor/examples/viewer_only/viewer_only_demo.tscn")
 @onready var _pretty_viewer_demo_scene: PackedScene = preload("res://addons/custom_graph_editor/examples/pretty_viewer/pretty_viewer_demo.tscn")
+@onready var _interactive_viewer_demo_scene: PackedScene = preload("res://addons/custom_graph_editor/examples/interactive_viewer/interactive_viewer_demo.tscn")
 
 
 func _ready() -> void:
@@ -27,6 +29,7 @@ func _ready() -> void:
     _inspector_action_example_button.pressed.connect(_on_inspector_action_button_pressed)
     _viewer_only_example_button.pressed.connect(_on_viewer_only_button_pressed)
     _pretty_viewer_example_button.pressed.connect(_on_pretty_viewer_button_pressed)
+    _interactive_viewer_example_button.pressed.connect(_on_interactive_viewer_button_pressed)
     _demo_mode_panel.visible = false
 
 
@@ -60,6 +63,10 @@ func _on_viewer_only_button_pressed() -> void:
 
 func _on_pretty_viewer_button_pressed() -> void:
     demo_editor(_pretty_viewer_demo_scene)
+
+
+func _on_interactive_viewer_button_pressed() -> void:
+    demo_editor(_interactive_viewer_demo_scene)
 
 
 # Show